backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
web3
requests
python-dotenv
sqlalchemy
psycopg2-binary
redis
cryptography
